+++ docs/@v2/changelog.md
@@ -7,6 +7,25 @@ toc:
 
 <!-- do-not-remove -->
 
+## 2.43.2 (2026-07-31)
+
+### Patch Changes
+
+- Updated @redocly/client-generator to v0.2.0.
+- Updated @redocly/openapi-core to v2.43.2.
+
+## 2.43.1 (2026-07-30)
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.43.1.
+
+## 2.43.0 (2026-07-30)
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.43.0.
+
 ## 2.42.0 (2026-07-30)
 
 ### Minor Changes