+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @redocly/cli
 
+## 2.43.2
+
+### Patch Changes
+
+- Updated @redocly/client-generator to v0.2.0.
+- Updated @redocly/openapi-core to v2.43.2.
+
+## 2.43.1
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.43.1.
+
+## 2.43.0
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.43.0.
+
 ## 2.42.0
 
 ### Minor Changes